+++ src/DTOs/Invoice/Attention.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Morningtrain\Economic\DTOs;
+namespace Morningtrain\Economic\DTOs\Invoice;
 
 use Morningtrain\Economic\Abstracts\Resource;
