+++ src/DTOs/Invoice/Recipient.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Morningtrain\Economic\DTOs;
+namespace Morningtrain\Economic\DTOs\Invoice;
 
 use Morningtrain\Economic\Abstracts\Resource;
 use Morningtrain\Economic\Enums\NemHandelType;
